[pyqgis] move qgis.core tricks to their own python file in additions folder
this makes the code much easier to read
